Refactor select and deselect frame
